V2.5.0 (#132)
支持多用户
支持SSL证书支持DNS验证
增加mysql测试用例
修复DNS设置时的展示歧义

+++ server/dto/tag.go
@@ -3,9 +3,9 @@ package dto
 import "encoding/json"
 
 type SearchTag struct {
-	Type    int `json:"type"`     // -1 不限
-	Status  int `json:"status"`   // -1 不限
-	GroupId int `json:"group_id"` // -1 不限
+	Type    int8 `json:"type"`     // -1 不限
+	Status  int8 `json:"status"`   // -1 不限
+	GroupId int  `json:"group_id"` // -1 不限
 }
 
 func (t SearchTag) ToString() string {
